requirements.txt
coverage==4.3.4
dj-database-url==0.4.2
Django==1.10.5
django-filter==1.0.1
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.4
django-recommends==0.4.0
mock==2.0.0
Markdown==2.6.8
psycopg2==2.6.2
pyfcm==1.2.8
gunicorn==19.6.0
whitenoise==3.3.0
